first

+++ study/studyapp/views.py
@@ -5,6 +5,9 @@ from django.db.models import Q
 from django.contrib import messages
 from django.contrib.auth import authenticate, login, logout
 from django.contrib.auth.models import User
+from django.contrib.auth.decorators import login_required
+from django.http import HttpResponse
+from django.contrib.auth.forms import UserCreationForm
 
 # rooms = [
 #     {'id': 1, 'name': 'Learn Python'},
@@ -15,8 +18,13 @@ from django.contrib.auth.models import User
 
 def login_page(request):
 
+    page = 'login'
+
+    if request.user.is_authenticated:
+        return redirect('home')
+
     if request.method == 'POST':
-        username = request.POST.get('username')
+        username = request.POST.get('username').lower()
         password = request.POST.get('password')
 
         try:
@@ -32,12 +40,28 @@ def login_page(request):
         else:
             messages.error(request, 'Username or password is incorrect')
 
-    context = {}
+    context = {'page': page}
     return render(request, 'studyapp/login_register.html', context)
 
 def logout_user(request):
     logout(request)
     return redirect('home')
+
+def register_user(request):
+    form = UserCreationForm()
+
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user  = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'An error occurred during registration')
+
+    return render(request, 'studyapp/login_register.html', {'form': form})
 
     
 
@@ -57,6 +81,7 @@ def room(request, pk):
     context = {'room': room}
     return render(request, 'studyapp/room.html', context)
 
+@login_required(login_url='login')
 def createroom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -67,9 +92,13 @@ def createroom(request):
     context = {'form': form}
     return render(request, 'studyapp/room_form.html', context)
 
+@login_required(login_url='login')
 def updateroom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse("You're not allowed here!")
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -79,8 +108,13 @@ def updateroom(request, pk):
     context = {'form': form}
     return render(request, 'studyapp/room_form.html', context)
 
+@login_required(login_url='login')
 def deleteroom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse("You're not allowed here!")
+    
     if request.method == 'POST':
         room.delete()
         return redirect('home')
